Do not use regex to parse /proc/config.gz.
During boot, a compatibility check is performed and
/proc/config.gz is parsed. /proc/config.gz has
a well defined format:

    # CONFIG_FOO is not set
    CONFIG_BAR=y

with no trailing / leading spaces, spaces around
'=', or trailing comments. Using regex is slow
and unnecessary, hence a normal string search
is performed instead.

On the other hand, for assemble_vintf that needs to parse
android-base.cfg, because the file is maintained by
humans, free format is accepted, and regex is used.

Test: boots (boot time compat check passes)
Test: boot time regression no longer happens (~120ms vs. ~90ms)
Test: libvintf_test
Bug: 63537988
Change-Id: If6ab207d40459689386aaf40cd835effc5597886