Load vintf manifests from the correct location.
Hardware differentiation is based on ro.product.device, not
ro.product.model. The vendor value of this
(ro.boot.product.hardware.sku) is used instead of ro.product.model so
that the correct value is used for GSI builds. After this change, the
odm manifest can be loaded from
/odm/etc/manifest_${ro.boot.product.hardware.sku}.xml or
/odm/etc/manifest.xml. Note that these files must be loaded from under
etc so that symlinks can be created on devices without an odm partition
as part of b/70678783. Moving the vendor manifest to etc is covered in
b/36790901, but is not required for correctness in this case because
the GSI is only ever used on devices with a vendor image.

Test: vintf_object_test
Bug: 70685002
Change-Id: Ie8c195fc844166ff24a40080e711860951afbe2a