Create libvintf_common
This is a static lib excluding utils.cpp. This
is originally libvintftest, but modified to
include RuntimeInfo and KernelConfigParser as well
for better generalization.

This lib can be used instead of libvintf when different
FileFetcher / PropertyFetcher / PartitionMounter
are needed, for example.

Test: builds
Test: libvintf_test
Test: vintf_object_test
Bug: 72722951

Change-Id: I3e4400bfb00bf0434a90a5fca9709bc76d5026c2
Merged-In: I3e4400bfb00bf0434a90a5fca9709bc76d5026c2