VintfObjectRecovery: mount all partitions in fstab
Specifically, ODM partition should be mounted when ODM manifest is loaded.

Previously, the on-device vendor manifest is not loaded correctly
because the ODM partition is never mounted. This can potentially
break framework-only OTAs.

For this to work, it requires a mount path for /odm to be in fstab.

If we are going to load VINTF metadata from partitions other than
/system, /vendor, /odm, the current code still works as long as the
mount path for /foo exists in fstab.

Test: manual OTA
Bug: 118635062
Change-Id: I9d02e36f262d4d0ca31f8b43dfe4486f13a2faa1